refactor: 회원가입 한 페이지로 리팩토링 (#43)
