+++ src/Router.tsx
@@ -18,8 +18,6 @@ import { SellerMypage } from 'pages/Seller/SellerMypage';
 import { SellerLetterWrite } from 'pages/Seller/SellerLetterWrite';
 import { Routes, Route, useNavigate } from 'react-router-dom';
 import { SellerMyPageReview } from 'pages/Seller/SellerMyPageReview';
-import { BuyerSignupPw } from 'pages/Buyer/BuyerSignupPw';
-import { BuyerSignupInfo } from 'pages/Buyer/BuyerSignupInfo';
 import { BuyerSignupComplete } from 'pages/Buyer/BuyerSignupComplete';
 import { BuyerFindInfo } from 'pages/Buyer/BuyerFindInfo';
 import { SellerVerifyMaterial } from 'pages/Seller/SellerVerifyMaterial';
@@ -83,8 +81,6 @@ const Router = () => {
       {/* 로그인 관련 페이지는 앞에 buyer seller 구분 제외했음 */}
       <Route path="/login" element={<BuyerLogin />} />
       <Route path="/signup" element={<BuyerSignup />} />
-      <Route path="/signup/setting" element={<BuyerSignupPw />} />
-      <Route path="/signup/info" element={<BuyerSignupInfo />} />
       <Route path="/signup/nav" element={<BuyerSignupComplete />} />
       <Route path="/find" element={<BuyerFindInfo />} />
       <Route path="/setting" element={<BuyerAccountSetting />} />
